refactor: reimpl using build-config
