+++ src/presets/aws-lambda/preset.ts
@@ -4,6 +4,16 @@ export type { AwsLambdaOptions as PresetOptions } from "./types";
 const awsLambda = defineNitroPreset(
   {
     entry: "./runtime/aws-lambda",
+    awsLambda: {
+      streaming: false,
+    },
+    hooks: {
+      "rollup:before": (nitro, rollupConfig) => {
+        if (nitro.options.awsLambda.streaming) {
+          (rollupConfig.input as string) += "-streaming";
+        }
+      },
+    },
   },
   {
     name: "aws-lambda" as const,
