+++ src/presets/aws-lambda/runtime/aws-lambda.ts
@@ -3,7 +3,6 @@ import type {
   APIGatewayProxyEventV2,
   APIGatewayProxyResult,
   APIGatewayProxyResultV2,
-  APIGatewayProxyStructuredResultV2,
   Context,
 } from "aws-lambda";
 import "#nitro-internal-pollyfills";
@@ -15,20 +14,18 @@ import {
   normalizeLambdaOutgoingHeaders,
 } from "nitropack/runtime/internal";
 import { withQuery } from "ufo";
-import { useRuntimeConfig } from "nitropack/runtime/config";
-import { Readable } from "node:stream";
 
 const nitroApp = useNitroApp();
 
-export async function bufferedHandler(
+export async function handler(
   event: APIGatewayProxyEvent,
   context: Context
 ): Promise<APIGatewayProxyResult>;
-export async function bufferedHandler(
+export async function handler(
   event: APIGatewayProxyEventV2,
   context: Context
 ): Promise<APIGatewayProxyResultV2>;
-export async function bufferedHandler(
+export async function handler(
   event: APIGatewayProxyEvent | APIGatewayProxyEventV2,
   context: Context
 ): Promise<APIGatewayProxyResult | APIGatewayProxyResultV2> {
@@ -81,95 +78,3 @@ export async function bufferedHandler(
     isBase64Encoded: awsBody.type === "binary",
   };
 }
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace awslambda {
-    // https://docs.aws.amazon.com/lambda/latest/dg/configuration-response-streaming.html
-    function streamifyResponse(
-      handler: (
-        event: APIGatewayProxyEventV2,
-        responseStream: NodeJS.WritableStream,
-        context: Context
-      ) => Promise<void>
-    ): any;
-
-    // eslint-disable-next-line @typescript-eslint/no-namespace
-    namespace HttpResponseStream {
-      function from(
-        stream: NodeJS.WritableStream,
-        metadata: {
-          statusCode: APIGatewayProxyStructuredResultV2["statusCode"];
-          headers: APIGatewayProxyStructuredResultV2["headers"];
-        }
-      ): NodeJS.WritableStream;
-    }
-  }
-}
-
-const streamingHandler = () =>
-  awslambda.streamifyResponse(
-    async (event: APIGatewayProxyEventV2, responseStream, context) => {
-      const query = {
-        ...event.queryStringParameters,
-      };
-      const url = withQuery(event.rawPath, query);
-      const method = event.requestContext?.http?.method || "get";
-
-      if ("cookies" in event && event.cookies) {
-        event.headers.cookie = event.cookies.join(";");
-      }
-
-      const r = await nitroApp.localCall({
-        event,
-        url,
-        context,
-        headers: normalizeLambdaIncomingHeaders(event.headers) as Record<
-          string,
-          string | string[]
-        >,
-        method,
-        query,
-        body: event.isBase64Encoded
-          ? Buffer.from(event.body || "", "base64").toString("utf8")
-          : event.body,
-      });
-      const httpResponseMetadata = {
-        statusCode: r.status,
-        headers: {
-          ...normalizeLambdaOutgoingHeaders(r.headers, true),
-          "Transfer-Encoding": "chunked",
-        },
-      };
-      if (r.body) {
-        const writer = awslambda.HttpResponseStream.from(
-          responseStream,
-          httpResponseMetadata
-        );
-        if (!r.body.getReader) {
-          writer.write(r.body);
-          writer.end();
-          return;
-        }
-        const reader = r.body.getReader();
-        await streamToNodeStream(reader, responseStream);
-        writer.end();
-      }
-    }
-  );
-
-const streamToNodeStream = async (
-  reader: Readable,
-  writer: NodeJS.WritableStream
-) => {
-  let readResult = await reader.read();
-  while (!readResult.done) {
-    writer.write(readResult.value);
-    readResult = await reader.read();
-  }
-  writer.end();
-};
-
-export const handler = useRuntimeConfig().streaming
-  ? streamingHandler()
-  : bufferedHandler;
